Introduce null type
Summary:
We failed to convince people that because `void` behaves at runtime as a singleton type inhabited by the value `null` it makes a perfect choice for the type of the expression `null`, so in this diff I introduce a new type `null` that will be used as the type of the expression  `null`.

The new type `null` is different from `void`, and the latter should eventually behave as an abstract type (e.g., a newtype wrapper around `null`).  In particular, we will allow `void` to instantiate generic type parameters, but will raise an error if one directly uses a value of type `void`.

This diff deals with the most straightforward stuff.  In particular, we remove the `void_is_type_of_null` flag and change a few places where `void` was used as the type of `null` to the new type `null`.

Reviewed By: CatherineGasnier

Differential Revision: D12825071

fbshipit-source-id: e84b4ffd44c58ac27b6da62b4f31fd139b826174