Expand trait decl in Hack
Summary:
This changes declaration to consider all trait methods at once instead of individually.

Previously, decl would build up a map of methods keyed by method name, and the one belonging to the last trait in the case of
```
class C {
  use T1, T2, T3;
}
```
would win. In the above example, assuming all traits declare method `f`, `T3::f` wins. This change makes it so that all the methods of the same name are first placed into a list until all traits are processed. Only then is the list collapsed subject to the previous rules.

The upside of this is that the lists can later be pruned, as in the case of method redeclaration, as opposed to the history of edits to the method map being lost.

Given
```
trait T1 {
  public function f(): void {}
  public function g(): void {}
}
trait T2 {
  public function e(): void {}
  public function f(): void {}
  public function h(): void {}
}

class C {
  use T1, T2;
}
```
before, methods were added to the decl record as `f g e T2::f h`, and now it is `e f T2::f g h`.

Reviewed By: jamesjwu

Differential Revision: D12896373

fbshipit-source-id: a139791567f62c24b049cd2f3ec578fbb750870f