Support metadata in backtrace_hash
Summary:
Some of the use-cases for backtrace_hash need to be aware of metadata fields.
In all the use-cases, we use backtrace_hash to memoize-gate the calls to backtrace: It's OK to get a false-positive where we get a NEW hash.
If a metadata is set (rare) or updated (even more rare) we should likely get a new hash

Reviewed By: markw65

Differential Revision: D10856409

fbshipit-source-id: 0962ae24f19ef2d0393015da245da5ff126a9b17