Reified generics in closures
Summary: This diff adds support for using reified generics in closures. The way it is done is by capturing all the reified generics that is used in the closure and assigning them to a local variable. When they are using inside the closure, we instead emit the temporary local variable.

Reviewed By: jamesjwu

Differential Revision: D12933286

fbshipit-source-id: eccaa10119f74dfa8c79e97651fcbadaf80992a5