Make null types print like option types
Summary:
Make nullable types that contain the type null appear as types the user would be able to denote, i.e. (T | null) prints as ?T.

Specifically, if null is the only type in an unresolved, just print "null"
If its in the unresolved with exactly one other type T, print "?T"
If it's with multiple, print "?(T1 | T2 ...)" etc.

Reviewed By: andrewjkennedy

Differential Revision: D12962983

fbshipit-source-id: e40194a5b80b572fb96659073be114c7c628eb1a