Forbid toplevel statements after an unbraced namespace declaration
Summary:
It looks like we neglected to scan items following an unbraced namespace declaration in the lowerer's `post_process` helper. This resulted in failing to emit errors for toplevel statements in files which have a namespace declaration with an empty body (like the added test case test_toplevel_stmt_in_namespace.php).

This change runs the `post_process` pass over the contents of the namespace, just as we do for a braced namespace declaration.

Reviewed By: kmeht

Differential Revision: D12932603

fbshipit-source-id: 38d34ee9b010f274f5d846726a7c3edbe9545e85