Add support for the null type in HHVM
Summary: D12825071 introduced a new type `null` such that the only value of type `null` is `null`.  The type `null` is different from `void`, which is also inhabited only by `null`, but apparently many people prefer to reserve `void` as a return type of functions that don't return any (useful) value.  HHVM already seems to have a notion of `null` type internally, and this diff exposes it, so that it is possible to use `null` in type hints and to access it through reflection and type structure mechanism.  We want to be able to distinguish `null` from `void` when accessed through reflection or type structure mechanism, which is why aliasing `null` to `void` doesn't work.

Reviewed By: ricklavoie

Differential Revision: D12922640

fbshipit-source-id: 614d298f25c459a4f8680350edec987a32d36490