Fix an inlining-related bug in FrameState
Summary:
In a couple of places in FrameState, we look up entries in the FPI stack to find
a matching entry.  This was done by comparing the returnSP and the returnSPOff.
However, the returnSPOff is relative to the frame pointer and, since inlining
changes the frame pointer, we could end up matching multiple entries in the FPI
stack.  This was causing an inlining bug because FrameState was annotating the
function information from AssertARFunc on the wrong FPI entry.  In debug builds,
this could trigger an assert, which this diff turns into always_assert.

Reviewed By: jano

Differential Revision: D12983096

fbshipit-source-id: 86dcf87c4177d58f341217b685f9fd4564f9fc7b