Enforce matching staticness for KeysetN key/val
Summary:
When we construct DArrLikeMap data for BKeyset types, we loosen
staticness for values but retain it for the TypedValue keys.  We then
convert these sometimes to DArrLikeMapN, which can result in a mismatch
in staticness between key and val.

checkInvariants() requires that we have exactly equal types between the
key and val, so let's unify it on construction.

Reviewed By: markw65

Differential Revision: D12969959

fbshipit-source-id: 6d92153f4c1d2928b4bb9876932f7b55f2010d00