Make getSortFunction less fragile.
Summary:
getSortFunction is fragile. It can clear the LSB of the enum value you pass in to it (if ascending was true), but it doesn't assert that you passed in something with the bit set. 2/3 call sites were passing values with the LSB cleared, so the function never did anything there, but it so happened that this worked since none of the escalateForSort methods currently care to distinguish between sort VS rsort or asort VS arsort, and the call site dealing with ksort/krsort correctly passed a value with the LSB set.
Make it better by asserting that the value passed in has the LSB cleared, and conditionally setting the LSB. The return value of getSortFunction now correctly distinguishes sort/rsort and asort/arsort (though we're still not making use of that distinction).

Reviewed By: markw65

Differential Revision: D12994848

fbshipit-source-id: 702846a9a2eb2d37b15bea38b1549a63d4cf4913