Deduplicate code and fix bugs in hack collection sorting.
Summary:
Sorting hack collections:
1) Duplicated a large quantity of the array sorting code.
2) All u*sorts were unsafe, because they would call mutate() to make sure the backing storage for the collection had exactly one ref and would sort that, ignoring the fact that the callback reenters and can observe the broken state of the hash during the sort or mutate the collection and cause the underlying storage to be reallocated (which it wouldn't notice).

Refactor hack collection sorting so that it uses the array sorting code to do 99% of the work.
Always copy the backing storage array when doing a u*sort on a collection so that the callback can't observe intermediate states during the sort, and can't cause the array being sorted to be reallocated.

Reviewed By: markw65

Differential Revision: D12994870

fbshipit-source-id: 1795569c95376442f8c738786b1edc40cdd53ea1