Add type variable environment for constraint solving
Summary:
We generalize the existing `tenv` and `subst` to a map from type variables to sets of lower and upper bounds.
The subtype simplifier generates a general constraint proposition; any top-level conjuncts in this proposition of the form `Tvar v <: t` or `t <: Tvar v` are put in the new type variable environment.
This is currently only performed when `--new-inference` is set.

Future tests should be placed in `test/typecheck/new_inference` that contains an appropriate `HH_FLAGS` file.

Reviewed By: manzyuk

Differential Revision: D12959987

fbshipit-source-id: 2ea33ac3fd4164abe57c8d2351f2c2050332270e