Temporarily forbid the use of the null type as a type hint
Summary: @[558351958:David] rightfully pointed out that even though I've added support for the `null` type in HHVM in D12922640, it's not going to be available for a couple of weeks until the next HHVM release, and during this time `null` used as a type hint will fatal in HHVM.  This diff temporarily hides `null` and makes it an internal only type.  Hopefully, once a new HHVM is released we will be able to simply back out this diff to enjoy full support for the `null` type.

Reviewed By: DavidSnider

Differential Revision: D12998815

fbshipit-source-id: 16c60f1432e4af7ad29f50cdb76f071c1eae220a