Selectively convert DecRef into DecRefNZ when destructors are disallowed
Summary:
When destructors are disallowed, we can make a DecRef cheaper by converting it
to a DecRefNZ, which avoids checking if the count got to zero and immediately
releasing the object. This will use more memory, but it is still OK to do
because the tracing GC can later collect the garbage produced.  This
transformation can also trigger more copy-on-write for arrays and strings for
which there are transitive references inside the object that is not immediately
released. To limit these overheads, this optimization is selectively performed based
on profiling and it's only applied when the number of times that the DecRef reached
0 is below a certain threshold.  When the type is know to be a string, we can be
more aggressive because they don't hold internal references to other objects, so
a separate threshold is used in that case.

Reviewed By: ricklavoie

Differential Revision: D12951309

fbshipit-source-id: 3182ee21e17453886c81e4328145ddc11cb86413