Add variance information to type variable environment
Summary:
For find "minimal" solutions to constraints, we will need to know the polarity (variance) of type variables with respect to the type of the expression being checked (e.g. the return type of a function call). Let's add `appears_covariantly` and `appears_contravariantly` fields to the type variable map, with this interpretation:

* `appears_covariantly = appears_contravariantly = true` is INVARIANT (appears in both positions)
* `appears_covariantly = appears_contravariantly = false` is sometimes called BIVARIANT (does not appear at all)
* Otherwise, it's purely covariant or purely contravariant.

This diff just sets up the map and provides some helpers.

Reviewed By: CatherineGasnier

Differential Revision: D12999893

fbshipit-source-id: c94d27f192bf95d6b9a11ba49e232d49c6a3c808