Don't generate redundant Tany constraints for parameter hints
Summary: When a function parameter has a default value, we check that its type is a subtype of the parameter type hint, and if there is no default value, we assume that there is one of type `Tany` and check that `Tany` is a subtype of the parameter type hint.  This is redundant and results in trivial constraints of the form `_ <: t` added to the current subtype proposition (stored as the field`subtype_prop` in the environment).

Reviewed By: andrewjkennedy

Differential Revision: D13021760

fbshipit-source-id: 46cfaa281c3bf412a91007631e3323c219511437