HHVM Debugger: Fix segfault from certain expression evaluations
Summary: Freeing the compilation unit right after the evaluation is causing segfaults in cases where a subsequent expression references a function, method or class defined by a previous expression. This diff changes the debugger to keep the units in memory until the request ends.  Unfortunately we have to keep them for the duration of the request, because there is no limitm to when the user in the debugger console can reference them again.

Reviewed By: markw65

Differential Revision: D12966534

fbshipit-source-id: eebbb52ca1a9bc75659405186e1031a3eb0b36d5