RFC: Disallow array access with invalid arraykey
Summary:
The following is legal to do in Hack.

```
$foo = darray[];
$foo[null] = 1; // $foo is now darray<?_, int>
```

However, at runtime, there's no such thing as a `darray<?_, int>`: `null` is cased to the empty string. We see a similar effect with `bool` and `float`.

For some construct:
```
X[Y] = Z;
```
If `X` is a `KeyedContainer<_, _>` add a constraint that `Y` must be an array key.

- This is put up for review without a flag, but the plumbing is in place to gate this
- I'm open to having this rule unconditionally applied to all `X[Y] = Z`

Reviewed By: jamesjwu

Differential Revision: D12902952

fbshipit-source-id: 42f770d0a9879e49fc8d0fae0d876f39cd5428f8