Add variance of type variables based on expression type
Summary:
We will be using the variance of a type variable wrt the type of an expression to "force" solutions to constraints. For example, in
```
function id<T>(T $x): T { return $x; }

function foo(C $c):void {
  $z = id($c);
  $z->bar();
}
```
it is possible to locally "solve" the constraint C <: v (where `v` is the type variable used for generic parameter `T`) by assuming `v = C`. This works because `T` appears only covariantly in the return type of `id`. If it had appeared contravariantly, or invariantly, as in
```
class Inv<T> {
  public function __construct(public T $item) { }
}
function make_inv<T>(T $x): Inv<T> { return new Inv($x); }
function foo(C $c):Inv<mixed> {
  $z = make_inv($c);
  return $z;
}
```
then we can't solve at the point of the `make_inv` call, because that would tie down the type of `$z` to be *exactly* `Inv<C>`.

To make all this work, we implement a function `set_tyvar_variance` that sets the "variance" info in the type variable environment according to where the type variables generated for generic parameters appear in the return type.

To see this, use `hh_show_env();` in the tests in `test/typecheck/new_inference`, which have examples such as the above.

Next step is to do local constraint solving, making use of the variance info. And also, add variance info anywhere that we generate fresh type variables, not just for function calls.

Reviewed By: manzyuk

Differential Revision: D13001388

fbshipit-source-id: 5576202b13d930f78f1e56d5ac8ccbfde2babe5c