Solve type variables by using their variance
Summary:
Use the variance information of fresh type variables to "solve" for a type variable on demand. For now, this means at call sites for method invocation.

The function `Typing_subtype.solve_tyvars ~tyvars env` will look at the variance and bounds on type variables `tyvars` in `env` to try and "solve" the constraints. For example:

```
function fst<T>(T $x, T $y): T {
  return $x;
}

function bar(C $c, D $d): void {
  // Here, let $x:v and C<:v and D<:v
  // Use the variance of v in the return type (covariant) to find v:=C|D
  $x = fst($c, $d);
  $x->foo();
}
```

Because `T` appears only covariantly in the return type of `fst`, we can "solve" the constraint `C, D <: v` to get `v := C | D`. This is the same technique used by Pierce and Turner in their work on Local Type Inference.

A couple of notes:
(1) With upper bounds (for type variables appearing only contravariantly), we currently only solve for singleton bounds. An empty bound could be assumed to be `mixed` - let's do that in a future diff, with an appropriate example. Multiple bounds would ideally correspond to an intersection type, which we don't support.
(2) With lower bounds, we currently only solve for non-empty sets. Ideally we would solve using an "empty" type, but unfortunately `Tunresolved []` is not interpreted properly as an empty type elsewhere in the type checker.

Reviewed By: CatherineGasnier

Differential Revision: D13024097

fbshipit-source-id: 5eb30bd8f17c8975ad4aedfe8939f36b3c7cab85