HHVM Debugger: Fix segfault when expanding closure in console REPL
Summary:
Currently if the user defines a closure in the console, we get back an object with a variablesReference so that it can be expanded. Trying to expand this segfaults the server because a closure is not actually a PHP/Hack object.

hphpd just prints the closure's dynamic class name (like 'Closure$;1234') with no children. It doesn't do this explicitly, but gets this behavior "for free" from the variables serializer.

Reviewed By: velocityboy

Differential Revision: D13054012

fbshipit-source-id: 15f8c14a2fcc605898851b6141c535f870ebd80b