Dedup types when printing and type-at-pos
Summary:
Deduplicate by using a typing set when printing the types from an unresolved. This gives better info and ignores things like (bool | bool) and returns just "bool" for type-at-pos.

Tests have changed because the order of the list is not preserved after using a set. Behavior has not changed at all.

Reviewed By: kmeht

Differential Revision: D12982213

fbshipit-source-id: e459a1e1772bd4d577293926b61d32e039c13b04