Use call-target profiling to enable more inlining of biased calls
Summary:
The existing CallTargetProfile already has information about the distribution of
the callees for each call site.  This diff leverages that profile information to
unlock more inlining opportunities.  If the call is biased towards a single
callee (beyond a certain threshold), a runtime check is emitted right before the
call.  If this check fails at runtime, a side exit is taken.  If the check
succeeds, the known callee information can enable inlining the callee.  In case
the known callee information does not trigger inlining, the runtime check is
removed to avoid its associated costs.

This new mechanism subsumes the previous "pgoPushedFunc" that was applied to
limited FPush* instructions.  This diff also removes that mechanism.

Reviewed By: ricklavoie

Differential Revision: D13074071

fbshipit-source-id: d16ffbdb998c5c6e5566480e6e7edf226f5f9b69