Don't redeclare files where changed typedefs were used when conservative_redecl is disabled
Summary: This is a case we missed in D9927186, because this conservative redeclaration logic lives in Decl_redecl_service rather than Decl_compare. As with class types, we need only recheck the files where the changed typedef was used, not redeclare them (If it were possible to extend a typedef, we would need to redeclare its extends dependencies, but Naming forbids this).

Reviewed By: dabek

Differential Revision: D13089486

fbshipit-source-id: 2f17d431aad8f1daf7de1af8b6f0934eba0de687