Fix pattern-matches on Tvar causing the type-checker to loop or throw under the new inference flag
Summary: I tried to run the type-checker tests with the `--new-inference` flag and found that the type-checker was hanging on [[ diffusion/FBS/browse/master/fbcode/hphp/hack/test/typecheck/tconst/generic_tc_invalid1.php | hphp/hack/test/typecheck/tconst/generic_tc_invalid1.php ]] and was also throwing an assert exception on some other test.  This diff puts in temporary fixes so that we can run the tests and track progress.  T36856670 serves as a reminder to do a proper work.

Reviewed By: CatherineGasnier

Differential Revision: D13084464

fbshipit-source-id: b68ed08800cb707404428a854f4f83735d5fa4b0