Solve type variables at the end of function or method definition
Summary: Forcing type variables to solutions after each function/method call is not enough.  In the examples below, there are no function/method calls in the body of the function/method, so we never force the type variables to solutions, and in particular we don't check if the upper bounds are compatible with the lower bounds.  We also need to resolve the type variables because they may be referenced by the types in the TAST.

Reviewed By: CatherineGasnier

Differential Revision: D13097549

fbshipit-source-id: 3b9d96b548532b2bb0ada93603d86fee777478aa