Make intish casts explicit when setting contents of superglobals
Summary:
This code is a mess and depdends heavily on PHP arrays. Lucky for us,
most of the actual setting-of-arrays occurs in register_variable which is
already explicitly casting array keys. Here's a few places that could
potentially intish cast. I'm not seeing logs for any of these but better safe
than sorry, I presume.

Reviewed By: mxw

Differential Revision: D13107675

fbshipit-source-id: 8fd9c3cba925a607f90b75ca8bb5f49812c6fe2a