Forbid truthiness tests on string and arraykey
Summary: Truthiness tests on values of type string or arraykey may not behave as expected. Like other dynamic languages, PHP considers `""` to be falsy, but unlike other dynamic languages, PHP considers `"0"` to be falsy. Hack has inherited this behavior. Since the user may have intended only to test for emptiness, we emit an error when testing the truthiness of a string after this change, suggesting `Str\is_empty` instead.

Reviewed By: CatherineGasnier

Differential Revision: D8877182

fbshipit-source-id: 862e9546b65d4f92a1056576f5c2b79cd0e7ede5