Introduce shallow_class declaration type
Summary:
We would like to introduce a class declaration type which contains only information which is syntactically available in the class definition. These shallow class declarations would consume less memory and be easier to compute and invalidate. In the future, we may build a "fast decl parser", which emits such a data structure directly from the parser using smart constructors.

This diff adds a simple `shallow_class` declaration type, which contains the subset of the `Nast.class_` fields which are used by the Decl module. It also changes the Decl module (and associated helper modules) to take a `shallow_class` as the input for class declaration rather than a `Nast.class_`. In future diffs, we can do more declaration work in Shallow_decl, and bring the `shallow_class` type closer to `decl_class_type`. As we do, we should be mindful of the fast-decl-parser use case--we should not perform any simplifications or error detection in Shallow_decl which could not be done in FFP smart constructors.

Reviewed By: dlreeves

Differential Revision: D12946876

fbshipit-source-id: b07a226ed4ea0452b7aee2748b014b6ecf5cbc44