a few corrections with Tany
Summary:
Under the flag --new-inference, we can safely say that for all T, T <: Tany and Tany <: T, so long we make sure (T | Tany) = (Tany | T) = T anywhere where we perform a union.

NB: This is not the case under the old inference because `expand_type ty` is Tany for a fresh type variable, and simplify_subtype does expand ty_sub and ty_super.

Reviewed By: andrewjkennedy

Differential Revision: D13070431

fbshipit-source-id: 39c4e8ecf2e1c84957727609cad587242e7d4b78