Fix some ubsan issues
Summary:
hhvm wouldn't even startup with ubsan enabled. It was mostly due to
things like &tv->m_data when tv was nullptr. This "worked" because
m_data is the first member, and the calling code checked the result
for nullptr (tv_val::is_set did exactly this check, for example).

Reviewed By: ricklavoie

Differential Revision: D13114528

fbshipit-source-id: f652a153273d578b165f5fe9ff95c99e6ec3bf21