Visibility for functions used in ptr arithmetic
Summary:
Even though doing pointer arithmetic is a bad idea because it is
breaking abstractions and making assumptions about code layout, we
can't always avoid it in a JIT and we need to be careful with this.
In an expression involving the address of a function, for example,
funcname - 5, the compiler should feed the linker with a relocation
against the original function symbol: the relocation target would be
funcname and addend would be minus 5. However, if this function
happens to be local, the relocation target will not use funcname
anymore, but it will be .text instead, and the addend will be a large
offset pointing to where funcname - 5 should be.  In this case, we
lose information required to reorder the the location of functions,
since funcname - 5 will not be valid anymore and we don't have the
original funcname reference to update it.

In HHVM, emitCallToExit() is one function that does pointer arithmetic
with assert_tc_saved_rip. With LTO, this function visibility changes
to local and the compiler toolchain will emit a relocation that lacks
a reference to the original symbol -- instead of
assert_tc_assert_rip-5, the relocation is .text+0xf4c3b00c, making it
impossible for BOLT to properly redo function layout without breaking
the expression involving function pointer arithmetic.

This diff changes the visibility of all functions whose address is
involved in function pointer arithmetic to global so their relocations
are emitted in a way that enables us to do post-link processing.

Reviewed By: maksfb

Differential Revision: D13111699

fbshipit-source-id: 5c5d7cfb6b2a6f7deba4a2e5bbbf3d1395a45b52