Handle the case when the subtype is a union in simplify_subtype under the new inference flag
Summary: What it says on the tin.  Effectively, this implements the rule that `(t1 | ... | tn) <: u` is equivalent to `t1 <: u & ... & tn <: u`.  This is true for unions, which is what we intend unresolved types to become in the new inference algorithm, but is not true for unresolved types as they are currently implemented.  (E.g., we treat the case of empty union on the left bound to a type variable specially.)  Hence, this change is guarded by the new inference flag.

Reviewed By: andrewjkennedy

Differential Revision: D13153680

fbshipit-source-id: 1607175f6074d2e4113b41337fd89989aa9c84de