Add typechecker test option to ignore error text
Summary:
When introducing fixes to the type checker it's common for error messages to change.
We introduce a new option `--ignore-error-text` on the test runner which has the effect of simply checking if both or neither output and expected match the text `No errors`.
Obviously this is very special-case and tied to the current behaviour of the type checker.

Reviewed By: CatherineGasnier

Differential Revision: D13176766

fbshipit-source-id: c44a7e85dffd34a45b3d1d411c9bcf788206af39