Replace hh_show by expect(...) in tests (1)
Summary:
The use of `hh_show` in tests is very brittle. It depends on pretty-printing; fixed points cause types to be shown more than once; and brittle position information is displayed.
Better to check that a variable has the expected type by passing it to an `expect(...)` function, which is very easy now that we have explicit generic parameters to functions.

Reviewed By: Wilfred

Differential Revision: D13178870

fbshipit-source-id: 2019c4558e157b8541ce903a15b14dc417de1d5f