Include top-level statements in the named and typed ASTs
Summary: This diff adds functionality to keep top level statements when converting an AST to an annotated AST (AAST) (e.g the named (NAST) and typed AST (TAST)). The reason for this change is that, in future, we would like to emit code using the annotated AST (e.g. NAST/TAST/etc.) so while top-level statements won't be type-checked, we still need them to be in the AAST so we can emit HHAS for them.

Reviewed By: Wilfred

Differential Revision: D12920458

fbshipit-source-id: 88a9746e3090704a0d75422bea7c910867f48c27