Make non-wildcard generics invalid in is-expressions
Summary:
Due to updates to the reified generics project, it should never be legal to use erased generics with the `is` operator. For example, `$x is Vector<string>` will **always** be false, because the operator will try to validate the type parameter (and Vector's type parameter isn't reified).

A followup diff will simplify the validation logic to be a simple option instead of its own variant.

Reviewed By: oulgen

Differential Revision: D13106441