Add more support for preserving aux bits
Summary:
Add support to ArgGroup for passing a typed-value to a function along
with a specified constant in the aux bits. Also add a new storeTVWithAux helper
function which writes a typed-value to memory along with a specified constant in
the aux bits. These aren't used yet, but will be used by the eager async memo
optimization.

Reviewed By: mxw

Differential Revision: D13071755

fbshipit-source-id: c7de2da0701634365e96c9e4c18459e743c6b1ae