Create `serverCommandLwt` module
Summary:
Linking against `Lwt_unix` causes `Unix.select` calls to fail with `EINTR`. Separate this out into a separate module so that we can link against it in the client (which uses Lwt throughout) and not the server (which uses Lwt nowhere, and would have failing `Unix.select` calls as a result).

In general, it's possible to recover from an `EINTR` by retrying the `select` (as `EINTR` simply means that the program received some signal, not that there was necessarily an error condition). However, we may end up in a situation where some amount of time was left to wait, and we would have to retry the `select` call with that amount of time remaining. Some Linux distributions may have `select` modify the `timeout` that was passed to it to indicate how much time was remaining, but the OCaml `Unix.select` does not do this, and I don't feel like it's very robust to calculate the remaining timeout manually.

Reviewed By: ljw1004

Differential Revision: D13197031

fbshipit-source-id: 4e93bf4445080ea6df8e25b16fa47dae3f5122b4