Speculate blocks to resolve branches (take three)
Summary:
If the next block has multiple predecessors, no side effects, and
multiple successors, we might be able to identify a single successor
block if we process it in conjunction with each predecessor, rather
than processing it using the union of the predecessor states as input.

eg

  if (($x ? 42 : 24) === 42) { ... }

can be reduced to

  if ($x) { ... }

The main change from the previous failed attempt is that this one
records the speculation results during analysis, and applies them
during optimize.

The issue with the original version is that optimize doesn't use
`groups` and so can get worse results from speculation than analysis
did, which can invalidate the block states.

Reviewed By: jano

Differential Revision: D13089598

fbshipit-source-id: 856d14f9ec3c0b9b3fbbe13c4399aebd9baf9394