Support null in is-expressions
Summary:
This backs out D12998815, adds an experimental flag for the `null` typehint, and adds type refinement for it via the `is` operator.

This one is a little different because for `null`, we refine in *both* branches of the condition. Additionally, in the "else" branch for an expression already typed as null, we refine to Tany, which is reasonable since that code won't be executed. If we ever have a true bottom type (empty union perhaps?) we should refine to that instead.

Reviewed By: manzyuk

Differential Revision: D13080629

fbshipit-source-id: b047a47a13dfe9adebe487a656b8b6bf3c48afd5