Implement as strict bytecode
Summary:
From Hack code (new syntax, parent diff)
```
trait T1 {
  public static async function f() {}
}

class C {
  use T1;
  final public static async function g() = T1::f;
}
```
we get bytecode of the form
```
.use T1 {
  T1::f as strict async [final public static] g;
  ...
}
```

First, remove `T1::f` in the preclass from the pair
```
{
  ...,
  "f" => [..., T1::f, ...],
  ...
}
```
and then delete all bindings for the new name, leaving only the declared one
```
{
  ...,
  "g" => [T1::f],
  ...
}
```

If the `async` and `static` modifiers are present, then they must match. If the trait method is `final`, then the redeclared method must be, but if the trait's method is not `final`, the redeclared method can be promoted to final.

Visibility promotions work both ways, i.e. `public` => `private` and vice-versa.

Reviewed By: markw65

Differential Revision: D12812251

fbshipit-source-id: b6198ae0d0dadb1895161d73d266f3a76b0b09be