Use select when pumping Watchman messages in Informant
Summary:
Since our entire existence is one big race condition with disk and Mercurial, it would be good to at least keep up in this race. When testing previous diffs, I've been seeing that various notifications are arriving suspiciously late (compared to, say, into Mercurial logs).

This is because the way we pump them is not reliable - there is only casual connection between `had_watchman_changes` variable and whether there actually are more Watchman messages. In practice this causes us to think that there are no more messages and going back to 1s sleep in `ServerMonitor.sleep_and_check`, while the messages are queueing.

Changing this logic to actually select on relevant FDs.

Reviewed By: 2BitSalute

Differential Revision: D13146321

fbshipit-source-id: f7ab1d07a9ad0404d49cb6c611dcc19f471f6ebb